[FXIOS-11351] Refactors, cleanup
